Create FindBookUseCaseTest and change Book's stock availability type to Long

+++ src/main/java/com/sansarch/bookstore_catalog_service/application/usecase/edit_book/dto/EditBookUseCaseInputBookDto.java
@@ -13,5 +13,5 @@ public class EditBookUseCaseInputBookDto {
     private String title;
     private String author;
     private BigDecimal price;
-    private Integer stockAvailability;
+    private Long stockAvailability;
 }
