+++ src/main/java/com/sansarch/bookstore_catalog_service/domain/book/exception/InsufficientStockException.java
@@ -7,7 +7,7 @@ import org.springframework.http.HttpStatus;
 public class InsufficientStockException extends RuntimeException {
     private final int status = HttpStatus.CONFLICT.value();
 
-    public InsufficientStockException(Long bookId, Integer available, Integer requested) {
+    public InsufficientStockException(Long bookId, Long available, Long requested) {
         super(String.format("Book ID %d: Requested %d exceeds available stock of %d.", bookId, requested, available));
     }
 }
